requirements.txt
# Зависимости для приложения «Расчёт трудоёмкости разработки ПС»

# GUI
PySide6>=6.5.0

# Экспорт
openpyxl>=3.1.0
python-docx>=1.0.0

# Диаграммы
matplotlib>=3.7.0

# Сборка (опционально, для создания exe)
pyinstaller>=6.0.0
